requirements.txt
rotate-screen
PyQt6-Qt6